Move legacy relation settings and entry title HTML into the adapter
