+++ src/FieldLayout/LayoutElements/Entries/EntryTitleField.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace CraftCms\Cms\FieldLayout\LayoutElements\Entries;
 
-use CraftCms\Cms\Cp\FormFields;
 use CraftCms\Cms\Element\Contracts\ElementInterface;
 use CraftCms\Cms\Entry\Elements\Entry;
 use CraftCms\Cms\Field\Enums\TranslationMethod;
@@ -87,39 +86,5 @@ class EntryTitleField extends TitleField
         }
 
         return $element->getType()->titleTranslationMethod->description();
-    }
-
-    #[Override]
-    public function inputHtml(?ElementInterface $element = null, bool $static = false): ?string
-    {
-        if (! $element instanceof Entry) {
-            throw new InvalidArgumentException(sprintf('%s can only be used in entry field layouts.', self::class));
-        }
-
-        $entryType = $element->getType();
-
-        if (! $entryType->hasTitleField) {
-            return null;
-        }
-
-        if ($entryType->allowLineBreaksInTitles) {
-            return FormFields::textareaHtml([
-                'class' => 'nicetext',
-                'id' => $this->id(),
-                'describedBy' => $this->describedBy($element, $static),
-                'rows' => 2,
-                'name' => $this->name ?? $this->attribute(),
-                'value' => $this->value($element),
-                'maxlength' => $this->maxlength,
-                'autofocus' => $this->autofocus,
-                'disabled' => $static || $this->disabled,
-                'readonly' => $this->readonly,
-                'required' => ! $static && $this->required,
-                'title' => $this->title,
-                'placeholder' => $this->placeholder,
-            ]);
-        }
-
-        return parent::inputHtml($element, $static);
     }
 }
